typescript(vx-demo): rewrite Treemap demo in TypeScript
